src/main/java/com/kernel360/kernelsquare/domain/tech_stack/service/TechStackService.java
package com.kernel360.kernelsquare.domain.tech_stack.service;

import com.kernel360.kernelsquare.domain.tech_stack.dto.CreateTechStackRequest;
import com.kernel360.kernelsquare.domain.tech_stack.dto.CreateTechStackResponse;
import com.kernel360.kernelsquare.domain.tech_stack.entity.TechStack;
import com.kernel360.kernelsquare.domain.tech_stack.repository.TechStackRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Service
@RequiredArgsConstructor
public class TechStackService {
    private final TechStackRepository techStackRepository;

    @Transactional
    public CreateTechStackResponse createTechStack(CreateTechStackRequest createTechStackRequest) {
        TechStack techStack = CreateTechStackRequest.toEntity(createTechStackRequest);
        techStackRepository.save(techStack);

        return CreateTechStackResponse.of(techStack);
    }
}
